Create COLONY
Bee colony algorithm

Collective intelligence is a branch of artificial intelligence based on the behavior of decentralized and self-organized systems.

- Using collective intelligence in solving problems

- Ant colony, bee colony and...

- Characteristics of collective intelligence

- Self-organized

- Division of work (people work together to achieve a final goal.)

- Getting an idea from the behavior of bees in solving problems

 

A colony of bees

The colony (collective behavior) of honey bees in nature consists of 3 parts:

- food sources

- The quality of the food source

- Ease of access

– Distance from the hive

– Worker bees

- They are related to the food source from which they are currently extracting nectar.

- Gather information about the distance, direction, profitability of the source and share the information in the hive

Some applications of this algorithm in engineering

– Neural network training for pattern recognition

- Scheduling work for production machines

- Categorization of information

- Optimizing the design of mechanical components

– Multiple optimizations

- Calibration of fuzzy logic controllers for athlete robots


The work mechanism in the hive is based on several principles; The Queen Bee, in these conditions, determines the platform of the founder role as the queen (king). Specify and create a proper way to upgrade and expand the network. Comprehensive and complete programs will be provided based on the engineering algorithm. In these conditions, the male players play the role of the bee, their duties will create coordination with the founder to create appropriate nodes and maintain the stability of the Queen (king). The worker's bee, according to the stretching form of the road map, the worker's bees include all our users around the world. Here we want to be stable in the security of the hive. The workers must be properly nourished in the pollination season. The network, based on the artificial intelligence algorithm, will have a sustainable security because our scale is widespread and wide. The chondo's duties are the tasks of maintenance and equipping. Each cell has a coordinates and has a owner. Policians are defined based on the encoding. The project's conditions create a strong and powerful reservoir with sufficient strength for all projects (Web3). These projects can have a high-power future against all factors that disrupt the growth of digital currency growth. We have disrupt agents We will donate a new insight into society. The best deserves best. In your opinion, what's the reason for the price of Bitcoin price at first to date. Only and only feelings of trust and ensure Satocosi Nakamoto to the community towards the creation of their work. Why not to date even a tradition has not been transferred from the property of Satoshi Nakamoto. Make my friend. The choice of this is with you if you want to trust bliss and join us.

»»If the white paper consent, the project will be loaded for community awareness and trust.««